+++ CHANGELOG.md
@@ -37,6 +37,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop support for TYPO3 9LTS (#1094)
 
 ### Fixed
+- Harden the `ConfigurationRegistry` (#1195)
 - Fix invalid array accesses in `TestingFramework` (#1179)
 - Fix invalid array accesses in `AbstractDataMapper` (#1170)
 - Only work on the auto increment value if the DB supports it (#1166)